Added eventLogger wizard
