+++ src/main/webapp/eventLogger/eventLogger.xhtml
@@ -1,0 +1,123 @@
+<!DOCTYPE html>
+<html xmlns="http://www.w3.org/1999/xhtml"
+      xmlns:h="http://java.sun.com/jsf/html"
+      xmlns:f="http://java.sun.com/jsf/core"
+      xmlns:p="http://primefaces.org/ui">
+
+<h:head>
+    <meta http-equiv="Content-Type" content="text/html; charset=utf-8"/>
+    <meta name="viewport" content="width=device-width, initial-scale=1.0"/>
+
+    <h:outputStylesheet name="css/app.css"/>
+
+    <link rel="stylesheet" href="//maxcdn.bootstrapcdn.com/font-awesome/4.3.0/css/font-awesome.min.css"/>
+</h:head>
+
+<h:body>
+    <h:form>
+        <p:growl id="growl" sticky="true" showDetail="true"/>
+
+        <p:wizard flowListener="#{eventWizard.onFlowProcess}" onnext="hideBackButton()">
+            <p:tab id="trackingTab" title="Tracking">
+                <p:panel header="Tracking">
+                    <p:messages />
+                    <h:panelGrid columns="1">
+                        <p:selectOneMenu id="trackingId" value="#{eventWizard.trackId}"
+                                         required="true" requiredMessage="Tracking missing">
+                            <f:selectItem itemLabel="Tracking" itemValue="" noSelectionOption="true"/>
+                            <f:selectItems value="#{eventWizard.trackIds}"/>
+                        </p:selectOneMenu>
+                    </h:panelGrid>
+                </p:panel>
+            </p:tab>
+
+            <p:tab id="portTab" title="Port">
+                <p:panel header="Port">
+                    <p:messages />
+                    <h:panelGrid columns="1">
+                        <p:selectOneMenu id="portId" value="#{eventWizard.location}"
+                                         required="true" requiredMessage="Port missing">
+                            <f:selectItem itemLabel="Port" itemValue="" noSelectionOption="true"/>
+                            <f:selectItems value="#{eventWizard.locations}"/>
+                        </p:selectOneMenu>
+                    </h:panelGrid>
+                </p:panel>
+            </p:tab>
+
+            <p:tab id="eventTab" title="Event">
+                <p:panel header="Event">
+                    <p:messages />
+                    <h:panelGrid columns="1">
+                        <p:selectOneMenu id="eventType" value="#{eventWizard.eventType}"
+                                         requiredMessage="Event missing" required="true">
+                            <f:selectItem itemLabel="Event" itemValue="" noSelectionOption="true"/>
+                            <f:selectItem itemLabel="RECEIVE" itemValue="RECEIVE"/>
+                            <f:selectItem itemLabel="LOAD" itemValue="LOAD"/>
+                            <f:selectItem itemLabel="UNLOAD" itemValue="UNLOAD"/>
+                            <f:selectItem itemLabel="CUSTOMS" itemValue="CUSTOMS"/>
+                            <f:selectItem itemLabel="CLAIM" itemValue="CLAIM"/>
+                        </p:selectOneMenu>
+                    </h:panelGrid>
+                </p:panel>
+            </p:tab>
+
+            <p:tab id="voyageTab" title="Voyage">
+                <p:panel header="Voyage">
+                    <p:messages />
+                    <h:panelGrid columns="1">
+                        <p:selectOneMenu id="voyage" value="#{eventWizard.voyageNumber}"
+                                         required="false" requiredMessage="Voyage missing">
+                            <f:selectItem itemLabel="Voyage" itemValue="#{eventBackingBean.voyageNumber}"/>
+                            <f:selectItems value="#{eventWizard.voyages}"/>
+                        </p:selectOneMenu>
+                    </h:panelGrid>
+                </p:panel>
+            </p:tab>
+
+            <p:tab id="dateTab" title="Completition Date">
+                <p:panel header="Voyage">
+                    <p:messages />
+                    <h:panelGrid columns="1">
+                        <p:calendar id="completionDate" value="#{eventWizard.completionDate}"
+                                    requiredMessage="A date is needed" required="false"
+                                    pattern="MM-dd-yyyy">
+                        </p:calendar>
+                    </h:panelGrid>
+                </p:panel>
+            </p:tab>
+
+            <p:tab id="confirm" title="Confirmation">
+                <p:panel header="Confirmation">
+                    <h:panelGrid id="confirmation" columns="3" columnClasses="grid,grid,grid">
+                        <h:panelGrid columns="2" columnClasses="label, value">
+                            <h:outputText value="Tracking id: " />
+                            <h:outputText value="#{eventWizard.trackId}" styleClass="outputLabel"/>
+
+                            <h:outputText value="Port : " />
+                            <h:outputText value="#{eventWizard.location}" styleClass="outputLabel"/>
+
+                            <h:outputText value="Event type: " />
+                            <h:outputText value="#{eventWizard.eventType}" styleClass="outputLabel"/>
+
+                            <h:outputText value="Voyage number: "/>
+                            <h:outputText value="#{eventWizard.voyageNumber}" styleClass="outputLabel"/>
+
+                            <h:outputText value="Date: " />
+                            <h:outputText value="#{eventWizard.completionDate}" styleClass="outputLabel"/>
+                        </h:panelGrid>
+                    </h:panelGrid>
+
+                    <p:commandButton value="Submit" action="#{eventWizard.save}" update="growl" process="@this"/>
+                </p:panel>
+            </p:tab>
+
+        </p:wizard>
+    </h:form>
+
+    <script>
+        function hideBackButton() {
+            $("div.ui-wizard-navbar>button.ui-wizard-nav-back").css("display", "none");
+        }
+    </script>
+</h:body>
+</html>
