+++ src/main/webapp/public/about.xhtml
@@ -1,0 +1,43 @@
+<html xmlns="http://www.w3.org/1999/xhtml"
+      xmlns:ui="http://xmlns.jcp.org/jsf/facelets">
+<ui:composition template="/WEB-INF/templates/common/public.xhtml">
+    <ui:define name="title">About</ui:define>
+    <ui:define name="content">
+        <div class="ui-grid ui-grid-responsive">
+            <div class="ui-grid-row">
+                <div class="ui-grid-col-3" id="leftCol"></div>
+                <div class="ui-grid-col-6" id="mainCol">
+                    <br/><br/>
+                    <div id="container" class="panel">
+                        <center>
+                            <a href="#{request.contextPath}"><img src="../images/CTlogobadge128.png"
+                                                                  style="padding-top: 10px;" height="96"/></a>
+                        </center>
+                        <h3>Tracking Your Cargo</h3>
+                        <p>The public tracking interface
+                            allows you to track your cargo and receive
+                            information such as current status, estimated
+                            time of arrival and
+                            detailed handling history.</p>
+                        <p>The interface does not require a login but you
+                            are expected to know your registered
+                            tracking ID - please contact your booking agent
+                            if you do not know your tracking ID. You should
+                            also contact your booking agent to make any
+                            modifications to your cargo such as a change in
+                            destination, a correction, etc or if your cargo
+                            is misdirected, delayed or held at customs.</p>
+                        <br/>
+                        <p>Visit <a
+                                href="https://github.com/eclipse-ee4j/cargotracker/">https://github.com/eclipse-ee4j/cargotracker/</a> for more information.
+                        </p>
+                        <hr/>
+                        <a href="track.xhtml"><i class="fa fa-arrow-left fa-lg" style="color: #999999;"></i></a>
+                    </div>
+                </div>
+                <div class="ui-grid-col-3" id="rightCol"></div>
+            </div>
+        </div>
+    </ui:define>
+</ui:composition>
+</html>
